Updated with auth controls

+++ EP-Flask-App/myhome/admin/views.py
@@ -9,25 +9,35 @@ admin_blueprints = Blueprint('admin', __name__, template_folder='templates/admin
 
 @admin_blueprints.route('/dashboard', methods=['GET', 'POST'])
 def dashboard():
+    if current_user.is_authenticated and current_user.role == 'admin':
+        user_id = current_user.id
+        print(f'User authenticated {current_user.username} and accessed dashboard in admin')
+        if request.method == 'GET':
+            print('GET Request to admin/dashboard')
+            users = User.query.all()    
 
-    user_id = current_user.id
-    if request.method == 'GET':
-        print('GET Request to admin/dashboard')
-        users = User.query.all()    
-
-    return render_template('usersdashboard.html', users=users)
+        return render_template('usersdashboard.html', users=users)
+    else:
+        print('User tried accessing admin.dashboard and not authorized')
+        return redirect(url_for('auth.login'))
 
 @admin_blueprints.route('/<int:user_id>/delete', methods=['GET', 'POST'])
 def delete(user_id):
 
-    cur_user_id = current_user.id
-    if cur_user_id == user_id:
-        users = User.query.all()  
-        return render_template('usersdashboard.html', users=users)
-    else:
-        print(f'Deleting user id {user_id}')
-        user = User.query.get_or_404(user_id)
-        db.session.delete(user)
-        db.session.commit()
+    if current_user.is_authenticated and current_user.role == 'admin':
+        cur_user_id = current_user.id
+        print(f'User authenticated {current_user.username} and deleting {user_id} in admin')
+        if cur_user_id == user_id:
+            print(f'User {current_user.username} cant delete itself')
+            users = User.query.all()  
+            return render_template('usersdashboard.html', users=users)
+        else:
+            print(f'Deleting user id {user_id}')
+            user = User.query.get_or_404(user_id)
+            db.session.delete(user)
+            db.session.commit()
 
-    return redirect(url_for('admin.dashboard'))
+        return redirect(request.referrer)
+    else:
+        print('User tried accessing admin.delete and not authorized')
+        return redirect(url_for('auth.login'))
